refactor(primitives): partial updates to newer maths & api
